api/requirements.txt
fastapi==0.109.2
python-multipart==0.0.9
uvicorn==0.27.1
opencv-python-headless==4.9.0.80
pillow==10.2.0
numpy==1.26.3